Add taus to the model state dict for EnsembleMapSaasSingleTaskGP (#3069)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/botorch/pull/3069

`EnsembleMapSaasSingleTaskGP` doesn't save the `taus` as part of the state dict. This is confusing as creating a new model followed by loading the state dict may result in the learned model hyperparameters being suboptimal and thus changing if calling `fit_gpytorch_mll` again. This diff adds `taus` as a buffer to avoid this issue.

Reviewed By: saitcakmak

Differential Revision: D85768264

fbshipit-source-id: f109329be6b02c3d29c689f5335eb2eafdcf483d