Use nanmean and nanstd in Standardize (#3072)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/botorch/pull/3072

Replaced `mean` and `std` with `nanmean` and `nanstd` (respectively) to improve robustness (and matching `StratifiedStandardize`). Also fixed a small bug in `nanstd` which caused the returned shapes to be incorrect if `keepdim=False`.

Note: The random seed in `test_transforms` in `test_fully_bayesian_multitask.py` was changed because the previous seed caused the test to fail on a single element due to randomness.

Reviewed By: dme65

Differential Revision: D86318672

fbshipit-source-id: b5af8d5f43a1f1f8e1233e2da2481127adc47242